Use gulp to output transpiled files to same directory as source files
The reason for doing this is to export Flow type definitions. When consuming
a library, Flow will look for files with the .js.flow extension with the same
name in the same directory as any .js file that is loaded. This lets consumers
use import the library without forcing a transpilation step; and it lets Flow
read type definitions from the original source files.